refactor: make test JVM constraints a Kotlin plugin
